Remove GAPScalaInteractiveShellLibrary.
