+++ examples/hoare/addition.scala
@@ -1,17 +1,19 @@
-import at.logic.gapt.cli.GAPScalaInteractiveShellLibrary._
 import at.logic.gapt.expr.Neg
-import at.logic.gapt.proofs.expansionTrees.METWeakQuantifier
+import at.logic.gapt.formats.hoare.ProgramParser
+import at.logic.gapt.proofs.expansionTrees.{compressQuantifiers, METWeakQuantifier}
 import at.logic.gapt.proofs.hoare.{ForLoop, SimpleLoopProblem}
 import at.logic.gapt.proofs.lk.LKToExpansionProof
+import at.logic.gapt.formats.prover9.Prover9TermParserLadrStyle._
+import at.logic.gapt.provers.prover9.Prover9Prover
 
-val p = parse.program("for y < z do x := s(x) od")
-val A = parse.p9("x = k")
-val B = parse.p9("x = k + z")
-val g_p0 = parse.p9("(all x (x + 0 = x))")
-val g_ps = parse.p9("(all x (all y (s(x+y) = x + s(y))))")
+val p = ProgramParser.parseProgram("for y < z do x := s(x) od")
+val A = parseFormula("x = k")
+val B = parseFormula("x = k + z")
+val g_p0 = parseFormula("(all x (x + 0 = x))")
+val g_ps = parseFormula("(all x (all y (s(x+y) = x + s(y))))")
 val g = List(g_p0, g_ps)
 
-val f = parse.p9("k + y = x")
+val f = parseFormula("k + y = x")
 
 val slp = SimpleLoopProblem(p.asInstanceOf[ForLoop], g, A, B)
 
@@ -23,10 +25,10 @@ println(slp.pi)
 
 val instanceSeq = slp.instanceSequent(2)
 println(instanceSeq)
-val proof = prover9.getProof(instanceSeq).get
+val proof = Prover9Prover.getLKProof(instanceSeq).get
 
 println( nLine + "Expansion sequent:")
-val expansionSequent = compressExpansionSequent(LKToExpansionProof(proof))
+val expansionSequent = compressQuantifiers(LKToExpansionProof(proof))
 expansionSequent.antecedent.foreach {
   case METWeakQuantifier(formula, instances) =>
     println(s"$formula:")
