+++ examples/hoare/array_init.scala
@@ -1,22 +1,25 @@
 import at.logic.gapt.expr.Neg
-import at.logic.gapt.proofs.expansionTrees.METWeakQuantifier
-import at.logic.gapt.cli.GAPScalaInteractiveShellLibrary._
+import at.logic.gapt.formats.hoare.ProgramParser
+import at.logic.gapt.formats.simple.SimpleFOLParser
+import at.logic.gapt.proofs.expansionTrees.{compressQuantifiers, METWeakQuantifier}
 import at.logic.gapt.proofs.hoare.{ForLoop, SimpleLoopProblem}
 import at.logic.gapt.proofs.lk.LKToExpansionProof
+import at.logic.gapt.formats.prover9.Prover9TermParserLadrStyle._
+import at.logic.gapt.provers.prover9.Prover9Prover
 
-val p = parse.program("for y < z do x := set(x, s(y), get(x, y)) od")
-val f = parse.p9("k <= z -> get(x,k) = get(x,0)")
-val g_s = parse.p9("(all x (s(x) != 0))")
-val g_lr = parse.p9("(all x (x <= x))")
-val g_0l = parse.p9("(all x (0 <= x))")
-val g_l0 = parse.p9("(all x (x <= 0 -> 0 = x))") // order of equality is important...
-val g_sl = parse.p9("(all x (all y (x <= y -> x <= s(y))))")
-val g_ls = parse.p9("(all x (all y (x <= s(y) -> x <= y | s(y) = x)))")
-val g_ge = parse.p9("(all x (all y (all z (get(set(x,y,z),y) = z))))")
-val g_gn = parse.p9("(all x (all y (all z (all w (w != y -> get(set(x,y,z),w) = get(x,w))))))")
+val p = ProgramParser.parseProgram("for y < z do x := set(x, s(y), get(x, y)) od")
+val f = parseFormula("k <= z -> get(x,k) = get(x,0)")
+val g_s = parseFormula("(all x (s(x) != 0))")
+val g_lr = parseFormula("(all x (x <= x))")
+val g_0l = parseFormula("(all x (0 <= x))")
+val g_l0 = parseFormula("(all x (x <= 0 -> 0 = x))") // order of equality is important...
+val g_sl = parseFormula("(all x (all y (x <= y -> x <= s(y))))")
+val g_ls = parseFormula("(all x (all y (x <= s(y) -> x <= y | s(y) = x)))")
+val g_ge = parseFormula("(all x (all y (all z (get(set(x,y,z),y) = z))))")
+val g_gn = parseFormula("(all x (all y (all z (all w (w != y -> get(set(x,y,z),w) = get(x,w))))))")
 val g = List(g_s, g_lr, g_0l, g_l0, g_sl, g_ls, g_ge, g_gn)
 
-val slp = SimpleLoopProblem(p.asInstanceOf[ForLoop], g, parse.fol("T()"), f)
+val slp = SimpleLoopProblem(p.asInstanceOf[ForLoop], g, SimpleFOLParser("T()"), f)
 
 println(slp.loop.body)
 println(slp.programVariables)
@@ -24,9 +27,9 @@ println(slp.pi)
 
 val instanceSeq = slp.instanceSequent(1)
 println(instanceSeq)
-val proof = prover9.getProof(instanceSeq).get
+val proof = Prover9Prover.getLKProof(instanceSeq).get
 
-val expansionSequent = compressExpansionSequent(LKToExpansionProof(proof))
+val expansionSequent = compressQuantifiers(LKToExpansionProof(proof))
 expansionSequent.antecedent.foreach {
   case METWeakQuantifier(formula, instances) =>
     println(s"$formula:")
