+++ examples/hol-tape/ntape-inst.scala
@@ -14,35 +14,43 @@ val filename = "examples/hol-tape/tape3-3.llk"
 
 /* begin of proof script  */
 
-import at.logic.gapt.cli.GAPScalaInteractiveShellLibrary._
-import at.logic.gapt.expr.fol.{undoHol2Fol, replaceAbstractions, reduceHolToFol}
+import java.nio.file.{Paths, Files}
+
 import at.logic.gapt.expr.hol._
 
-import at.logic.gapt.expr.fol.undoHol2Fol
+import at.logic.gapt.formats.arithmetic.HOLTermArithmeticalExporter
+import at.logic.gapt.formats.hlk.HLKHOLParser
+import at.logic.gapt.formats.latex.SequentsListLatexExporter
 
 import at.logic.gapt.formats.llk.HybridLatexParser
 import at.logic.gapt.algorithms.rewriting.DefinitionElimination
+import at.logic.gapt.formats.tptp.TPTPHOLExporter
+import at.logic.gapt.formats.writers.FileWriter
 import at.logic.gapt.proofs.HOLSequent
 import at.logic.gapt.proofs.lk.base.HOLSequentOrdering
-import at.logic.gapt.proofs.lk.{AtomicExpansion, regularize, subsumedClausesRemovalHOL}
-import at.logic.gapt.proofs.lksk.sequentToLabelledSequent
-import at.logic.gapt.proofs.resolution.RobinsonToRal
+import at.logic.gapt.proofs.lk._
 
-import at.logic.gapt.provers.prover9._
 import at.logic.gapt.proofs.ceres.clauseSets._
-import at.logic.gapt.proofs.ceres.projections.Projections
 import at.logic.gapt.proofs.ceres.struct.StructCreators
 
-import at.logic.gapt.proofs.ceres.ceres_omega
-import at.logic.gapt.proofs.lksk.LKskToExpansionProof
-import at.logic.gapt.proofs.lk.LKToLKsk
+object exportLatex {
+  def apply( list: List[HOLSequent], fn: String ) = {
+    val writer = ( new FileWriter( fn ) with SequentsListLatexExporter with HOLTermArithmeticalExporter )
+    writer.exportSequentList( list, Nil ).close
+  }
+}
+
+object exportTHF {
+  def apply( ls: List[HOLSequent], filename: String, positive: Boolean = false ) =
+    Files.write(Paths.get(filename), TPTPHOLExporter( ls, positive ).getBytes)
+}
 
  val nLine = sys.props("line.separator")
 
  def show(s:String) = println( nLine + nLine + "+++++++++ "+s+" ++++++++++" + nLine )
 
  show("Loading file")
- val pdb = loadLLK(filename)
+ val pdb = HybridLatexParser(filename)
 
  show("Eliminating definitions, expanding tautological axioms")
  val elp = AtomicExpansion(DefinitionElimination(pdb.Definitions, regularize(pdb.proof("TAPEPROOF"))))
@@ -56,10 +64,10 @@ import at.logic.gapt.proofs.lk.LKToLKsk
  show("Computing clause set")
  //val cl = AlternativeStandardClauseSet(struct)
  val cl = SimpleStandardClauseSet(struct)
- //val cl = structToClausesList(struct).map(_.toHOLSequent)
+ //val cl = StandardClauseSet.transformStructToClauseSet(struct).map(_.toHOLSequent)
 
  show("simplifying clause set")
- val fs = HOLSequent(Nil, List(parseLLKFormula("var x :i; x = x")))
+ val fs = HOLSequent(Nil, List(HLKHOLParser.parseFormula("var x :i; x = x")))
  val rcl = subsumedClausesRemovalHOL(fs::(cl.toList))
  val tcl = deleteTautologies(rcl)
