+++ examples/hol-tape/ntape-script.scala
@@ -16,9 +16,8 @@ val filename = "./examples/hol-tape/ntape.llk"
 
 /* begin of proof script  */
 
-import at.logic.gapt.cli.GAPScalaInteractiveShellLibrary.loadLLK
 import at.logic.gapt.expr._
-import at.logic.gapt.expr.fol.{undoHol2Fol, replaceAbstractions, reduceHolToFol}
+import at.logic.gapt.expr.fol.{replaceAbstractions, reduceHolToFol}
 import at.logic.gapt.expr.hol._
 
 import at.logic.gapt.expr.fol.undoHol2Fol
@@ -27,7 +26,7 @@ import at.logic.gapt.formats.llk.HybridLatexParser
 import at.logic.gapt.algorithms.rewriting.DefinitionElimination
 import at.logic.gapt.proofs.FOLClause
 import at.logic.gapt.proofs.lk.base._
-import at.logic.gapt.proofs.lk.{AtomicExpansion, regularize, LKToLKsk}
+import at.logic.gapt.proofs.lk.{AtomicExpansion, regularize}
 import at.logic.gapt.proofs.lksk.sequentToLabelledSequent
 import at.logic.gapt.proofs.resolution.RobinsonToRal
 
@@ -78,7 +77,7 @@ import at.logic.gapt.proofs.lk.LKToLKsk
   }
 
       show("Loading file")
-      val pdb = loadLLK(filename)
+      val pdb = HybridLatexParser(filename)
 
       show("Eliminating definitions, expanding tautological axioms")
       val elp = AtomicExpansion(DefinitionElimination(pdb.Definitions, regularize(pdb.proof("TAPEPROOF"))))
