+++ src/main/scala/at/logic/gapt/cli/CLIMain.scala
@@ -21,21 +21,31 @@ object CLIMain {
 """
 
   val imports = """
-  import at.logic.gapt.cli.GAPScalaInteractiveShellLibrary._
+  import at.logic.gapt.examples._
   import at.logic.gapt.expr._
   import at.logic.gapt.expr.fol._
   import at.logic.gapt.expr.hol._
+  import at.logic.gapt.formats.simple._
+  import at.logic.gapt.formats.prover9._
+  import at.logic.gapt.formats.hlk._
+  import at.logic.gapt.formats.xml._
+  import at.logic.gapt.formats.llk._
   import at.logic.gapt.proofs.expansionTrees._
   import at.logic.gapt.proofs.hoare._
   import at.logic.gapt.proofs._
+  import at.logic.gapt.proofs.ceres._
+  import at.logic.gapt.proofs.ceres.clauseSets._
   import at.logic.gapt.proofs.lk._
   import at.logic.gapt.proofs.lk.base._
+  import at.logic.gapt.proofs.lk.cutIntroduction._
   import at.logic.gapt.proofs.lk.cutIntroduction.Deltas._
   import at.logic.gapt.proofs.lksk
   import at.logic.gapt.proofs.occurrences.FormulaOccurrence
-  import at.logic.gapt.provers.minisat.MiniSATProver
-  import at.logic.gapt.prooftool.{Main => PT}
-  import help.{apply => help}
+  import at.logic.gapt.provers.minisat._
+  import at.logic.gapt.provers.prover9._
+  import at.logic.gapt.provers.veriT._
+  import at.logic.gapt.prooftool.prooftool
+  import at.logic.gapt.utils._
   import at.logic.gapt.cli.GPL.{apply => copying, printLicense => license}
   """
