+++ src/main/scala/at/logic/gapt/formats/hlk/HOLParser.scala
@@ -251,7 +251,7 @@ class HLKHOLParser {
 
   }
 
-  def parseFormula( s: String ): LambdaExpression = f( parse( s ) )
+  def parseFormula( s: String ): HOLFormula = f( parse( s ) )
 
 }
