+++ src/main/scala/at/logic/gapt/formats/veriT/VeriTParser.scala
@@ -4,7 +4,7 @@ import at.logic.gapt.expr.BetaReduction._
 import at.logic.gapt.expr._
 import at.logic.gapt.expr.fol.{ FOLMatchingAlgorithm, FOLSubstitution }
 import at.logic.gapt.expr.hol.{ removeQuantifiers, instantiate, isPrenex }
-import at.logic.gapt.proofs.expansionTrees.{ formulaToExpansionTree, ExpansionTree, ETWeakQuantifier, ExpansionSequent }
+import at.logic.gapt.proofs.expansionTrees._
 import java.io.{ Reader, FileReader }
 import scala.collection.immutable.HashMap
 import scala.util.parsing.combinator._
@@ -231,6 +231,9 @@ object VeriTParser extends RegexParsers {
   def getExpansionProof( filename: String ): Option[ExpansionSequent] = {
     getExpansionProof( new FileReader( filename ) )
   }
+
+  def getExpansionProofWithSymmetry( fileName: String ): Option[ExpansionSequent] =
+    getExpansionProof( fileName ) map { addSymmetry( _ ) }
 
   // NOTE: The expansion proof returned is a tautology modulo symmetry!!!!
   def getExpansionProof( reader: Reader ): Option[ExpansionSequent] = {
