+++ src/test/scala/at/logic/gapt/proofs/ceres/ceres_omegaTest.scala
@@ -1,7 +1,10 @@
 package at.logic.gapt.proofs.ceres
 
-import at.logic.gapt.cli.GAPScalaInteractiveShellLibrary._
+import at.logic.gapt.algorithms.rewriting.DefinitionElimination
 import at.logic.gapt.expr._
+import at.logic.gapt.formats.llk.HybridLatexParser
+import at.logic.gapt.proofs.ceres.projections.Projections
+import at.logic.gapt.proofs.ceres.struct.StructCreators
 import at.logic.gapt.proofs.{ Ant, Suc, HOLSequent }
 import at.logic.gapt.proofs.lk._
 import at.logic.gapt.proofs.lk.base._
@@ -18,12 +21,12 @@ import at.logic.gapt.proofs.ceres.clauseSets._
 class ceres_omegaTest extends Specification with ClasspathFileCopier {
 
   def prepareProof( file: String, proofname: String ) = {
-    val p = loadLLK( tempCopyOfClasspathFile( file ) )
-    val elp = definitionElimination( p.Definitions, regularize( p.proof( proofname ) ) )
+    val p = HybridLatexParser( tempCopyOfClasspathFile( file ) )
+    val elp = AtomicExpansion( DefinitionElimination( p.Definitions, regularize( p.proof( proofname ) ) ) )
     val selp = LKToLKsk( elp )
-    val struct = extractStruct( selp )
+    val struct = StructCreators.extract( selp )
     val ls = SimpleStandardClauseSet( struct )
-    val proj = computeProjections( selp )
+    val proj = Projections( selp )
     ( selp, ls, struct, proj )
   }
 
@@ -69,7 +72,7 @@ class ceres_omegaTest extends Specification with ClasspathFileCopier {
       val ( p, cs, struct, proj ) = prepareProof( "llk/simple-leibnizeq.llk", "THEPROOF" )
       val rp = refutation1( cs )
 
-      val ( acnf, _ ) = CERESomega( proj, rp, sequentToLabelledSequent( p.root ), struct )
+      val ( acnf, _ ) = ceres_omega( proj, rp, sequentToLabelledSequent( p.root ), struct )
       val et = LKskToExpansionProof( acnf )
       ok
     }
